API: add locked files

+++ apps/api/src/contracts/workspace/deleteWorkspaceNode.ts
@@ -1,5 +1,6 @@
 import { S } from 'schema';
 import { WorkspaceNodeCollection } from '../../collections/WorkspaceNode';
+import { AppError } from '../../common/errors';
 import { findNodeAllChildren } from '../../common/workspace-tree';
 import { createContract, createGraphqlBinding } from '../../lib';
 import { getNodeByIdWithCheck } from './_common';
@@ -15,6 +16,9 @@ export const deleteWorkspaceNode = createContract(
   .fn(async (appUser, id) => {
     const node = await getNodeByIdWithCheck(appUser, id);
     const removeNodes = [node, ...(await findNodeAllChildren(node._id))];
+    if (removeNodes.some(x => x.isLocked)) {
+      throw new AppError('Cannot remove locked node');
+    }
     await WorkspaceNodeCollection.deleteMany({
       _id: {
         $in: removeNodes.map(x => x._id),
